Updated refactored code
The files have been renamed as part of the refactoring.
