+++ src/components/Main/QuizConfigurator.js
@@ -1,0 +1,171 @@
+import React, { useState } from 'react';
+import PropTypes from 'prop-types';
+import {
+  Container,
+  Segment,
+  Item,
+  Dropdown,
+  Divider,
+  Button,
+  Message,
+} from 'semantic-ui-react';
+
+import mindImg from '../../images/mind.svg';
+
+import {
+  CATEGORIES,
+  NUM_OF_QUESTIONS,
+  DIFFICULTY,
+  QUESTIONS_TYPE,
+  COUNTDOWN_TIME,
+} from '../../constants';
+import { shuffle } from '../../utils';
+
+import Offline from '../Offline';
+
+const Main = ({ startQuiz }) => {
+  const [category, setCategory] = useState('0');
+  const [numOfQuestions, setNumOfQuestions] = useState(5);
+  const [difficulty, setDifficulty] = useState('easy');
+  const [questionsType, setQuestionsType] = useState('0');
+  const [countdownTime, setCountdownTime] = useState({
+    hours: 0,
+    minutes: 120,
+    seconds: 0,
+  });
+  const [processing, setProcessing] = useState(false);
+  const [error, setError] = useState(null);
+  const [offline, setOffline] = useState(false);
+
+  const handleTimeChange = (e, { name, value }) => {
+    setCountdownTime({ ...countdownTime, [name]: value });
+  };
+
+  const renderContent = () => (
+    <Container>
+      <Segment>
+        <Item.Group divided>
+          <Item>
+            <Item.Image src={mindImg} />
+            <Item.Content>
+              <Item.Header>
+                <h1>The Ultimate Trivia Quiz</h1>
+              </Item.Header>
+              {error && (
+                // Comment: Display error message if there's an error.
+                <Message error onDismiss={() => setError(null)}>
+                  <Message.Header>Error!</Message.Header>
+                  {error.message}
+                </Message>
+              )}
+              <Divider />
+              <Item.Meta>
+                {/* Comment: Dropdowns for selecting quiz parameters */}
+                <p>In which category do you want to play the quiz?</p>
+                <Dropdown
+                  fluid
+                  selection
+                  name="category"
+                  placeholder="Select Quiz Category"
+                  header="Select Quiz Category"
+                  options={CATEGORIES}
+                  value={category}
+                  onChange={(e, { value }) => setCategory(value)}
+                  disabled={processing}
+                />
+                {/* ... Other dropdowns ... */}
+              </Item.Meta>
+              <Divider />
+              <Item.Extra>
+                <Button
+                  primary
+                  size="big"
+                  icon="play"
+                  labelPosition="left"
+                  content={processing ? 'Processing...' : 'Play Now'}
+                  onClick={fetchData}
+                  disabled={!allFieldsSelected || processing}
+                />
+              </Item.Extra>
+            </Item.Content>
+          </Item>
+        </Item.Group>
+      </Segment>
+      <br />
+    </Container>
+  );
+
+  //Function to fetch quiz data from an API
+  const fetchData = () => {
+    setProcessing(true);
+
+    if (error) setError(null);
+
+    const API = `https://opentdb.com/api.php?amount=${numOfQuestions}&category=${category}&difficulty=${difficulty}&type=${questionsType}`;
+
+    fetch(API)
+      .then(response => response.json())
+      .then(data =>
+        setTimeout(() => {
+          const { response_code, results } = data;
+
+          if (response_code === 1) {
+            // Handle insufficient questions error
+            const message = (
+              <p>
+                The API doesn't have enough questions for your query. (Ex.
+                Asking for 50 Questions in a Category that only has 20.)
+                <br />
+                <br />
+                Please change the <strong>No. of Questions</strong>,{' '}
+                <strong>Difficulty Level</strong>, or{' '}
+                <strong>Type of Questions</strong>.
+              </p>
+            );
+
+            setProcessing(false);
+            setError({ message });
+
+            return;
+          }
+
+          // Comment: Shuffle options for each question
+          results.forEach(element => {
+            element.options = shuffle([
+              element.correct_answer,
+              ...element.incorrect_answers,
+            ]);
+          });
+
+          setProcessing(false);
+          // Start the quiz with fetched data and countdown time
+          startQuiz(
+            results,
+            countdownTime.hours + countdownTime.minutes + countdownTime.seconds
+          );
+        }, 1000)
+      )
+      .catch(error =>
+        setTimeout(() => {
+          if (!navigator.onLine) {
+            // Comment: Handle offline scenario
+            setOffline(true);
+          } else {
+            setProcessing(false);
+            setError(error);
+          }
+        }, 1000)
+      );
+  };
+
+  // Render Offline component if offline
+  if (offline) return <Offline />;
+
+  return <Container>{renderContent()}</Container>;
+};
+
+Main.propTypes = {
+  startQuiz: PropTypes.func.isRequired,
+};
+
+export default Main;
