+++ src/components/Result/QNA.js
@@ -2,9 +2,12 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { Table } from 'semantic-ui-react';
 
+// Component for displaying Questions and Answers table
 const QNA = ({ questionsAndAnswers }) => {
-  return (
+  // Render the Q&A table
+  const renderQnaTable = () => (
     <Table celled striped selectable size="large">
+      {/* Table Header */}
       <Table.Header>
         <Table.Row>
           <Table.HeaderCell>No.</Table.HeaderCell>
@@ -14,8 +17,10 @@ const QNA = ({ questionsAndAnswers }) => {
           <Table.HeaderCell>Points</Table.HeaderCell>
         </Table.Row>
       </Table.Header>
+      {/* Table Body */}
       <Table.Body>
         {questionsAndAnswers.map((item, i) => (
+          // Row for each question and answer set
           <Table.Row key={i + 1}>
             <Table.Cell>{i + 1}</Table.Cell>
             <Table.Cell>{item.question}</Table.Cell>
@@ -27,8 +32,11 @@ const QNA = ({ questionsAndAnswers }) => {
       </Table.Body>
     </Table>
   );
+
+  return renderQnaTable();
 };
 
+// PropTypes for type-checking
 QNA.propTypes = {
   questionsAndAnswers: PropTypes.array.isRequired,
 };
