+++ src/components/Result/QuizResult.js
@@ -5,6 +5,7 @@ import { Container, Menu } from 'semantic-ui-react';
 import Stats from './Stats';
 import QNA from './QNA';
 
+// Result component displays either Stats or QNA based on the selected tab
 const Result = ({
   totalQuestions,
   correctAnswers,
@@ -13,14 +14,40 @@ const Result = ({
   replayQuiz,
   resetQuiz,
 }) => {
+  // State to track the active tab (default: 'Stats')
   const [activeTab, setActiveTab] = useState('Stats');
 
+  // Handle tab click event to switch between 'Stats' and 'QNA'
   const handleTabClick = (e, { name }) => {
     setActiveTab(name);
   };
 
+  // Function to render the content based on the active tab
+  const renderContent = () => {
+    if (activeTab === 'Stats') {
+      return (
+        <Stats
+          totalQuestions={totalQuestions}
+          correctAnswers={correctAnswers}
+          timeTaken={timeTaken}
+          replayQuiz={replayQuiz}
+          resetQuiz={resetQuiz}
+        />
+      );
+    }
+
+    if (activeTab === 'QNA') {
+      return <QNA questionsAndAnswers={questionsAndAnswers} />;
+    }
+
+    // Add additional cases for other tabs if needed
+    return null;
+  };
+
+  // Render the Result component
   return (
     <Container>
+      {/* Menu for switching between 'Stats' and 'QNA' tabs */}
       <Menu fluid widths={2}>
         <Menu.Item
           name="Stats"
@@ -33,21 +60,17 @@ const Result = ({
           onClick={handleTabClick}
         />
       </Menu>
-      {activeTab === 'Stats' && (
-        <Stats
-          totalQuestions={totalQuestions}
-          correctAnswers={correctAnswers}
-          timeTaken={timeTaken}
-          replayQuiz={replayQuiz}
-          resetQuiz={resetQuiz}
-        />
-      )}
-      {activeTab === 'QNA' && <QNA questionsAndAnswers={questionsAndAnswers} />}
+
+      {/* Render content based on the active tab */}
+      {renderContent()}
+      
+      {/* Add some spacing */}
       <br />
     </Container>
   );
 };
 
+// PropTypes for type-checking
 Result.propTypes = {
   totalQuestions: PropTypes.number.isRequired,
   correctAnswers: PropTypes.number.isRequired,
@@ -57,4 +80,5 @@ Result.propTypes = {
   resetQuiz: PropTypes.func.isRequired,
 };
 
+// Export the Result component
 export default Result;
